+++ src/excel/excel_functions/unicode.rb
@@ -1,0 +1,10 @@
+module ExcelFunctions
+  
+  def unicode(a)
+    return a if a.is_a?(Symbol)
+    return :value if a == nil
+    a = string_argument(a)
+    return a.ord
+  end
+  
+end